requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
apipkg==3.0.1
astor==0.8.1
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
colorama==0.4.6
configparser==5.3.0
crayons==0.4.0
execnet==1.9.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.2.0
pytest-ast-transformer==1.0.3
pytest-forked==1.4.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-soft-assertions==0.1.2
pytest-xdist==3.0.2
requests==2.28.1
selenium==4.6.0
six==1.16.0
texttable==1.6.7
toml==0.10.2
urllib3==1.26.13
webdriver-manager==3.8.5
openpyxl==3.0.10
